split up shared choice card components
