+++ packages/modules/src/modules/banners/choiceCardsBanner/components/ChoiceCards.tsx
@@ -5,8 +5,9 @@ import { until } from '@guardian/src-foundations/mq';
 import { visuallyHidden } from '@guardian/src-foundations/accessibility';
 import { HasBeenSeen, useHasBeenSeen } from '../../../../hooks/useHasBeenSeen';
 import { contributionType, ChoiceCardProps } from '../../../shared/helpers/choiceCards';
-import { ChoiceCardAmountButtons, ChoiceCardFrequencyTabs } from '../../../shared/ChoiceCard';
 import { space } from '@guardian/src-foundations';
+import { ChoiceCardAmountButtons } from '../../../shared/choiceCard/ChoiceCardAmountButtons';
+import { ChoiceCardFrequencyTabs } from '../../../shared/choiceCard/ChoiceCardFrequencyTabs';
 
 const styles = {
     epicFrequenciesGroupOverrides: css`
